backend/requirements.txt
black==24.10.0
coverage==7.6.12
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
flake8==7.1.1
isort==5.13.2
pylint==3.3.3
python-dotenv==1.0.1
whitenoise==6.9.0